Add more data fetchers

+++ WhatTheShop/Program.cs
@@ -1,7 +1,6 @@
 ﻿using ClosedXML.Excel;
 using Microsoft.EntityFrameworkCore;
 using WhatTheShop.DB;
-using WhatTheShop.Models;
 using WhatTheShop.Services;
 
 namespace WhatTheShop;
@@ -29,19 +28,16 @@ internal class Program
         //await RunAnalyticsVisit(apiClient, wtsToken, zones, devices);
         //await RunAnalyticsRecord(apiClient, wtsToken, zones, devices);
         //await RunAnalyticsInstore(apiClient, wtsToken, zones, devices);
-        await RunAnalyticZones(apiClient, db);
-        //await RunAnalyticsRaw(apiClient, wtsToken, zones, devices);
+        //await RunAnalyticZones(apiClient, db);
+        await RunAnalyticRaw(apiClient, db);
         //await RunAnalyticDevice(apiClient, db);
-        //await RunAnalyticsSystem(apiClient, wtsToken, zones, devices);
-        //await RunMediaVisits(apiClient, wtsToken, zones, devices);
-        //await RunMediaCampaign(apiClient, wtsToken, zones, devices);
-        //await RunAnalyticSensor(apiClient, wtsToken, zones, devices);
+        //await RunAnalyticSystem(apiClient, db);
+        // await RunMediaVisits(apiClient, db);
+        //await RunMediaCampaign(apiClient, db);
+        //await RunAnalyticSensor(apiClient, db);
 
         // save excel file
-        if (File.Exists(fileName))
-        {
-            File.Delete(fileName);
-        }
+        if (File.Exists(fileName)) File.Delete(fileName);
 
         workbook.SaveAs(fileName);
 
@@ -51,7 +47,7 @@ internal class Program
 
     private static async Task RunUser(ApiClient apiClient, DbCtx db)
     {
-        var service = new UserService(apiClient, db, OverwriteDb);
+        var service = new UserService(apiClient, db);
 
         await service.FetchZones();
         await service.FetchZoneInfos();
@@ -98,29 +94,15 @@ internal class Program
         await service.FetchAnalyticVisitorCountCommon();
     }
 
-     private static async Task RunAnalyticSensor(ApiClient apiClient, List<Zone> zones, List<Device> devices)
+    private static async Task RunAnalyticRaw(ApiClient apiClient, DbCtx db)
     {
-        Console.WriteLine("Fetching /1/analytic/sensor/count...");
-        var analyticSensorCount = await apiClient.GetAnalyticSensorCount(zones);
-    }
+        var service = new AnalyticRawService(apiClient, db);
 
-    private static async Task RunMediaCampaign(ApiClient apiClient, List<Zone> zones)
-    {
-        Console.WriteLine("Fetching /1/media/campaign/list...");
-        var campaignList = await apiClient.GetCampaignList(zones);
-
-        Console.WriteLine("Fetching  /1/media/campaign/count...");
-        var campaignCount = await apiClient.GetCampaignCount(zones, campaignList);
-    }
-
-    private static async Task RunMediaVisits(ApiClient apiClient)
-    {
-
-    }
-
-    private static async Task RunAnalyticsSystem(ApiClient apiClient)
-    {
-
+        await service.FetchAnalyticRawServicePasserby();
+        await service.FetchAnalyticRawServiceVisitor();
+        await service.FetchAnalyticRawServiceVisitorLight();
+        await service.FetchAnalyticRawServiceVisitorMacList();
+        await service.FetchAnalyticRawServicePasserbyMacList();
     }
 
     private static async Task RunAnalyticDevice(ApiClient apiClient, DbCtx db)
@@ -130,14 +112,13 @@ internal class Program
         await service.FetchAnalyticDeviceCount();
     }
 
-    private static async Task RunAnalyticsRaw(ApiClient apiClient)
+    private static async Task RunAnalyticZones(ApiClient apiClient, DbCtx db)
     {
+        var service = new AnalyticZoneService(apiClient, db);
 
-    }
-
-    private static async Task RunAnalyticZones(ApiClient apiClient)
-    {
-
+        await service.FetchAnalyticZonesGeneral();
+        await service.FetchAnalyticZonesVenn();
+        // await service.FetchAnalyticZonesSankey(); // always empty
     }
 
     private static async Task RunAnalyticsInstore(ApiClient apiClient)
@@ -150,5 +131,45 @@ internal class Program
 
     private static async Task RunAnalyticsVisit(ApiClient apiClient)
     {
+    }
+
+    private static async Task RunAnalyticSystem(ApiClient apiClient, DbCtx db)
+    {
+        var service = new AnalyticSystemService(apiClient, db);
+
+        await service.FetchAnalyticSystemLastUpdate();
+        await service.FetchAnalyticSystemQuickLastUpdate();
+        await service.FetchAnalyticSystemForceRefresh();
+        await service.FetchAnalyticSystemTemporaryTable();
+    }
+
+    private static async Task RunMediaVisits(ApiClient apiClient, DbCtx db)
+    {
+        var service = new MediaVisitService(apiClient, db);
+
+        await service.FetchMediaVisitCount();
+        await service.FetchMediaVisitCountDetails();
+    }
+
+    private static async Task RunMediaCampaign(ApiClient apiClient, DbCtx dbCtx)
+    {
+        Console.WriteLine("Fetching /1/media/campaign/list...");
+        var campaignList = await apiClient.GetCampaignList(dbCtx.Zones.ToList()); // always empty
+
+        Console.WriteLine("Fetching  /1/media/campaign/count...");
+        // requires campaigns so always empty for now
+        var campaignCount = await apiClient.GetCampaignCount(dbCtx.Zones.ToList(), campaignList);
+    }
+
+    private static async Task RunAnalyticSensor(ApiClient apiClient, DbCtx db)
+    {
+        var service = new AnalyticSensorService(apiClient, db);
+
+        await service.FetchAnalyticSensorCount();
+        await service.FetchAnalyticSensorCountDetails();
+
+        // it requires sensor or hardware id. It's not available in the API
+        // await service.FetchAnalyticSensorCountRaw();  // here
+        // await service.FetchAnalyticSensorCountDetailsRaw(); // same here
     }
 }
